+++ Config/formatters.php
@@ -1,0 +1,5 @@
+<?php return [
+    'html'  => new \Kadet\Highlighter\Formatter\HtmlFormatter(),
+    'cli'   => new \Kadet\Highlighter\Formatter\CliFormatter(),
+    'debug' => new \Kadet\Highlighter\Formatter\DebugFormatter()
+];