+++ Language/Java.php
@@ -42,4 +42,12 @@ class Java extends CSharp // evil
     {
         return 'java';
     }
+
+    public static function getAliases()
+    {
+        return [
+            'name'      => ['java'],
+            'mime'      => ['text/x-java'],
+            'extension' => ['*.java']        ];
+    }
 }